Bugfix 7397 vaccination status not updated (#7466)
* #7397 - Cleanup

* #7397 - Refactoring

* #7397 - Refactoring
